Consolidated EvidenceUpload tests on single file
